software/pyproject.toml
[tool.poetry]
name = "playt-player"
version = "0.1.0"
description = "A clean architecture audio player for physical media cartridges"
authors = ["OpenPlayt Team"]
readme = "README.md"
packages = [{include = "playt_player"}]

[tool.poetry.dependencies]
python = "^3.11"
python-vlc = "^3.0.20123"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
ruff = "^0.1.0"
mypy = "^1.6.0"
types-python-vlc = "^0.1.0"

[tool.poetry.scripts]
playt = "playt_player.interface.cli.player_cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "vlc.*",
    "pydub.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
]


